Add support for parallel polling and job processing...
Using the maxBatchSize property the worker can query for batches of jobs, and as those jobs are worked on it will continue to query for more jobs as slots become available.

Also moves error handling code for a failed job into its own bean so implementations can replace the bean for custom error handling.